Core/Social: fix uint back to 32 because of low

+++ src/server/game/Entities/Player/SocialMgr.cpp
@@ -324,7 +324,7 @@ void SocialMgr::BroadcastToFriendListers(Player* player, WorldPacket* packet)
 
     TeamId teamId = player->GetTeamId();
     AccountTypes security = player->GetSession()->GetSecurity();
-    uint64 guid = player->GetGUIDLow();
+    uint32 guid = player->GetGUIDLow();
     bool allowTwoSideWhoList = sGameConfig->GetBoolConfig("AllowTwoSide.WhoList");
     AccountTypes gmLevelInWhoList = AccountTypes(sGameConfig->GetIntConfig("GM.InWhoList.Level"));
